절대 경로 추출 방법 변경

+++ phantom/modules/capture.initialize.js
@@ -8,8 +8,10 @@ var initialize = function () {
 	//Promise 객체를 리턴
 	return new Promise(function (resolve, reject) {
 		try {
-			//output 폴더를 삭제 (PhantomJS의 removeTree 함수는 동기 방식으로 동작)
-			fs.removeTree(vars.outputFilePath);
+			if (fs.exists(vars.outputFilePath))) {
+				//output 폴더를 삭제 (PhantomJS의 removeTree 함수는 동기 방식으로 동작)
+				fs.removeTree(vars.outputFilePath);
+			}
 
 			//완료
 			resolve(true);
